Finaly!!!!!!!!!!!!!
Everything works as it should. The problem was in my logic :D

I thought that React Component 's state, that directly connected to props, will update automaticaly when the props are updated. But i was wrong. I had to use componentDidUpdate, when props change and rewrite state on my own.

Searching for the solution, I rewrote all the components to elements. The code is too dirty, but it works and the shots save after we click "End Turn" button.

That's all for today!