Changed ship creation. Now we have createShip function, that accepts length of the desired ship. Then it defines type of the ship on its own and returns a ship wrapped in AddDrag
Also, changed canDropShip and placeShip functions that we can place our ship on both fields(our and opponents). It determines needed battlefield depending on player and places ship there.

Henceforth we can't "place"(we couldn't do it earlier, but we could hover opponents field and drop a ship there it would place on that coordinates but on our field) ship where we're supposed to shoot and we can't shoot where we're supposed to place our ship(we clicked on a cell and it would be marked as shot at other field)
I changed parameters which are passed to fillWithCells function, when we creating  PlayersBattleField and OpponentsBattleField.
Heretofore we passed onClick to field with ships and canDrop+placeShip to shots field. We pass empty functions instead of them.

Minor css changes