Made new drag preview for ships. Now it's sticked to our pointer.
Also I've got a bug there if we move pointer to close to the right edge of the window -- our preview image starts to shrink.

And there's a real problem with perfomance. After preview image change we've got a slideshow with ~10fps. Gotta take care of it.

Added parseShipName function in helpers.js